Moved util to own module.
