+++ src/io/parquet/read/primitive/mod.rs
@@ -1,4 +1,5 @@
-use std::{convert::TryInto, hint::unreachable_unchecked};
+mod utils;
+use utils::ExactChunksIter;
 
 use parquet2::{
     encoding::{bitpacking, hybrid_rle, uleb128, Encoding},
@@ -6,7 +7,7 @@ use parquet2::{
     types::NativeType,
 };
 
-use super::utils;
+use super::utils as other_utils;
 use super::{ColumnChunkMetaData, ColumnDescriptor};
 use crate::{
     array::PrimitiveArray,
@@ -14,48 +15,8 @@ use crate::{
     buffer::MutableBuffer,
     datatypes::DataType,
     error::{ArrowError, Result},
-    trusted_len::TrustedLen,
     types::NativeType as ArrowNativeType,
 };
-
-struct ExactChunksIter<'a, T: NativeType> {
-    chunks: std::slice::ChunksExact<'a, u8>,
-    phantom: std::marker::PhantomData<T>,
-}
-
-impl<'a, T: NativeType> ExactChunksIter<'a, T> {
-    #[inline]
-    pub fn new(slice: &'a [u8]) -> Self {
-        assert_eq!(slice.len() % std::mem::size_of::<T>(), 0);
-        let chunks = slice.chunks_exact(std::mem::size_of::<T>());
-        Self {
-            chunks,
-            phantom: std::marker::PhantomData,
-        }
-    }
-}
-
-impl<'a, T: NativeType> Iterator for ExactChunksIter<'a, T> {
-    type Item = T;
-
-    #[inline]
-    fn next(&mut self) -> Option<Self::Item> {
-        self.chunks.next().map(|chunk| {
-            let chunk: <T as NativeType>::Bytes = match chunk.try_into() {
-                Ok(v) => v,
-                Err(_) => unsafe { unreachable_unchecked() },
-            };
-            T::from_le_bytes(chunk)
-        })
-    }
-
-    #[inline]
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        self.chunks.size_hint()
-    }
-}
-
-unsafe impl<'a, T: NativeType> TrustedLen for ExactChunksIter<'a, T> {}
 
 fn read_dict_buffer_optional<T, A, F>(
     buffer: &[u8],
@@ -265,7 +226,7 @@ where
                     read_required(page.buffer(), additional, values, op)
                 }
                 _ => {
-                    return Err(utils::not_implemented(
+                    return Err(other_utils::not_implemented(
                         &page.encoding(),
                         is_optional,
                         page.dictionary_page().is_some(),
@@ -301,7 +262,7 @@ where
                 read_required::<T, A, F>(page.buffer(), additional, values, op)
             }
             _ => {
-                return Err(utils::not_implemented(
+                return Err(other_utils::not_implemented(
                     &page.encoding(),
                     is_optional,
                     page.dictionary_page().is_some(),
